test_models.py

import google.generativeai as genai
import os
import time
from dotenv import load_dotenv

load_dotenv()
genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

models_to_test = [
    "models/gemini-1.5-flash",
    "models/gemini-1.5-flash-001",
    "models/gemini-1.5-flash-latest",
    "models/gemini-1.5-pro",
    "models/gemini-1.5-pro-001",
    "models/gemini-2.0-flash",
    "models/gemini-2.0-flash-lite",
    "models/gemini-2.0-flash-001",
    "models/gemini-2.0-pro-exp-02-05", # emerging models
]

print("Testing models for generation access...")

for model_name in models_to_test:
    print(f"\nTesting {model_name}...")
    try:
        model = genai.GenerativeModel(model_name)
        response = model.generate_content("Hello")
        print(f"✅ SUCCESS: {model_name}")
        print(f"Response: {response.text}")
        break # Found a working one!
    except Exception as e:
        print(f"❌ FAILED: {model_name}")
        print(f"Error: {e}")
        time.sleep(1) # Be nice to the API
