requirements.txt
# Core Dependencies
crawl4ai>=0.3.0
supabase>=2.0.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Data Processing
markdownify>=0.11.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
tenacity>=8.2.0
rich>=13.0.0
tqdm>=4.66.0
aiohttp>=3.9.0
numpy>=1.24.0
tiktoken>=0.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
